Split project in 4 libs: * common (common parts for all other libs) * analyser * corrector * tools * Fix #26
