+++ src/SConscript
@@ -1,6 +1,17 @@
 Import( 'env' )
 Import( 'loudnessAssessmentVersionStr' )
 
+### loudness Common ###
+
+loudnessCommonLibStatic = env.StaticLibrary(
+    'loudnessCommon',
+    Glob( 'loudnessCommon/*.cpp' ),
+)
+
+env.Alias( 'install', env.Install( 'lib', loudnessCommonLibStatic ) )
+
+Export( 'loudnessCommonLibStatic' )
+
 ### loudness Analyser ###
 
 # check boost accumulators (header only library)
@@ -18,12 +29,14 @@ loudnessAnalyserLib = env.SharedLibrary(
     loudnessAnalyserLibName,
     Glob( 'loudnessAnalyser/*.cpp' ) +
     Glob( 'loudnessAnalyser/utils/*.cpp' ),
+    LIBS = [loudnessCommonLibStatic]
 )
 
 loudnessAnalyserLibStatic = env.StaticLibrary(
     'loudnessAnalyser',
     Glob( 'loudnessAnalyser/*.cpp' ) +
     Glob( 'loudnessAnalyser/utils/*.cpp' ),
+    LIBS = [loudnessCommonLibStatic]
 )
 
 # Install loudnessAnalyser libs
@@ -34,6 +47,36 @@ env.Alias( 'install', env.Install( 'lib', loudnessAnalyserLibStatic ) )
 env.Alias( 'install', env.Install( 'include/loudnessAnalyser', "loudnessAnalyser/LoudnessAnalyser.hpp" ) )
 
 Export( 'loudnessAnalyserLibStatic' )
+
+### loudness Corrector ###
+
+loudnessCorrectorLibName = 'loudnessCorrector'
+# add version number to the name of the shared library if Windows platform
+if env['PLATFORM'] == 'win32':
+    loudnessCorrectorLibName += '-' + loudnessAssessmentVersionStr
+loudnessCorrectorLib = env.SharedLibrary(
+    loudnessCorrectorLibName,
+    Glob( 'loudnessCorrector/*.cpp' ),
+    LIBS = [
+        loudnessCommonLibStatic,
+        loudnessAnalyserLibStatic,
+    ]
+)
+
+loudnessCorrectorLibStatic = env.StaticLibrary(
+    'loudnessCorrector',
+    Glob( 'loudnessCorrector/*.cpp' ),
+    LIBS = [
+        loudnessCommonLibStatic,
+        loudnessAnalyserLibStatic,
+    ]
+)
+
+# Install loudnessCorrector libs
+env.Alias( 'install', env.InstallVersionedLib( 'lib', loudnessCorrectorLib ) )
+env.Alias( 'install', env.Install( 'lib', loudnessCorrectorLibStatic ) )
+
+Export( 'loudnessCorrectorLibStatic' )
 
 ### loudness Tools ###
 
@@ -54,8 +97,10 @@ if conf.CheckLibWithHeader(sndfileLib, 'sndfile.h', 'c'):
         loudnessToolsLibName += '-' + loudnessAssessmentVersionStr
     loudnessToolsLib = sndfileEnv.SharedLibrary(
         loudnessToolsLibName,
-        Glob( 'tool/*/*.cpp' ),
+        Glob( 'loudnessTools/*.cpp' ) +
+        Glob( 'loudnessTools/io/*.cpp' ),
         LIBS = [
+            loudnessCommonLibStatic,
             loudnessAnalyserLibStatic,
             sndfileLib,
         ],
@@ -63,8 +108,10 @@ if conf.CheckLibWithHeader(sndfileLib, 'sndfile.h', 'c'):
 
     loudnessToolsLibStatic = sndfileEnv.StaticLibrary(
         'loudnessTools',
-        Glob( 'tool/*/*.cpp' ),
+        Glob( 'loudnessTools/*.cpp' ) +
+        Glob( 'loudnessTools/io/*.cpp' ),
         LIBS = [
+            loudnessCommonLibStatic,
             loudnessAnalyserLibStatic,
             sndfileLib,
         ],
