+++ src/loudnessAnalyser/Loudness.cpp
@@ -1,5 +1,4 @@
 #include "Loudness.hpp"
-#include "common.hpp"
 #include <numeric>
 #include <cmath>
