+++ src/loudnessAnalyser/Loudness.hpp
@@ -2,14 +2,16 @@
 #ifndef _LOUDNESS_ANALYSER_LOUDNESS_HPP_
 #define _LOUDNESS_ANALYSER_LOUDNESS_HPP_
 
-#include <vector>
+#include <loudnessCommon/common.hpp>
+#include "Histogram.hpp"
+#include "LoudnessAnalyser.hpp"
+
 #include <boost/accumulators/accumulators.hpp>
 #include <boost/accumulators/statistics/stats.hpp>
 #include <boost/accumulators/statistics/rolling_mean.hpp>
 #include <boost/accumulators/statistics/rolling_mean.hpp>
 
-#include "Histogram.hpp"
-#include "LoudnessAnalyser.hpp"
+#include <vector>
 
 using namespace boost::accumulators;
