+++ src/loudnessAnalyser/LoudnessAnalyser.hpp
@@ -2,7 +2,7 @@
 #ifndef _LOUDNESS_ANALYSER_LOUDNESS_ANALYSER_HPP_
 #define _LOUDNESS_ANALYSER_LOUDNESS_ANALYSER_HPP_
 
-#include "common.hpp"
+#include <loudnessCommon/common.hpp>
 
 #include <cstdlib>
 #include <vector>
