+++ src/loudnessAnalyser/Process.cpp
@@ -1,6 +1,4 @@
-
 #include "Process.hpp"
-#include "common.hpp"
 
 namespace Loudness{
