+++ src/loudnessAnalyser/TruePeakMeter.hpp
@@ -1,8 +1,7 @@
-
-#include <vector>
-
 #ifndef _LOUDNESS_ANALYSER_TRUE_PEAK_METER_HPP_
 #define _LOUDNESS_ANALYSER_TRUE_PEAK_METER_HPP_
+
+#include <loudnessCommon/common.hpp>
 
 #include <cstdlib>
 #include <vector>
