+++ src/loudnessCommon/HardwareDetection.cpp
