+++ src/loudnessCommon/HardwareDetection.hpp
