+++ src/loudnessCorrector/LookAheadLimiter.hpp
@@ -1,11 +1,14 @@
 #ifndef LOOK_AHEAD_LIMITER_H
 #define LOOK_AHEAD_LIMITER_H
 
+#include <loudnessCommon/common.hpp>
+
+#include "RollingMax.hpp"
+
 #include <boost/accumulators/accumulators.hpp>
 #include <boost/accumulators/statistics/stats.hpp>
 #include <boost/accumulators/statistics/rolling_sum.hpp>
 #include <boost/circular_buffer.hpp>
-#include "RollingMax.hpp"
 
 #include <iostream>
 #include <cmath>
@@ -15,26 +18,26 @@ using namespace boost::accumulators;
 namespace Loudness {
 namespace tool {
 
-class LookAheadLimiter
+class LoudnessExport LookAheadLimiter
 {
 public:
 	LookAheadLimiter( const float lookAheadTime, const float sampleRate, const float threshold );
 	~LookAheadLimiter();
-	
+
 	bool process( float &value );
 	bool getLastSamples( float &value );
-	
+
 private:
 	size_t signalCircularBufferSize;
 	size_t maxCircularBufferSize;
-	
+
 	boost::circular_buffer<float> signal;
-	
+
 	RollingMax<float> accMax;
-	
+
 	accumulator_set<float, stats<tag::rolling_sum> > accShortTimeSum;
 	accumulator_set<float, stats<tag::rolling_sum> > accShortTimeSum2;
-	
+
 	float threshold;
 	size_t processTime;
 	size_t lastSamples;
